+++ packages/react/src/components/controls/MediaDeviceSelect.tsx
@@ -61,6 +61,7 @@ export const MediaDeviceSelect: (
     ref,
   ) {
     const room = useMaybeRoomContext();
+    const previousActiveDeviceId = React.useRef<string>('default');
     const handleError = React.useCallback(
       (e: Error) => {
         if (room) {
@@ -91,14 +92,15 @@ export const MediaDeviceSelect: (
     }, [onDeviceListChange, devices]);
 
     React.useEffect(() => {
-      if (activeDeviceId && activeDeviceId !== '') {
+      if (activeDeviceId !== previousActiveDeviceId.current) {
         onActiveDeviceChange?.(activeDeviceId);
       }
+      previousActiveDeviceId.current = activeDeviceId;
     }, [activeDeviceId]);
 
     const handleActiveDeviceChange = async (deviceId: string) => {
       try {
-        await setActiveMediaDevice(deviceId, { exact: exactMatch });
+        await setActiveMediaDevice(deviceId, { exact: exactMatch ?? true });
       } catch (e) {
         if (e instanceof Error) {
           onDeviceSelectError?.(e);