+++ packages/react/src/hooks/index.ts
@@ -55,3 +55,5 @@ export * from './useTrackTranscription';
 export * from './useVoiceAssistant';
 export * from './useParticipantAttributes';
 export * from './useIsRecording';
+export * from './useTextStream';
+export * from './useTranscriptions';